Refactor polyfills to be in a single closure
This refactors the two polyfills to share a single closure that wraps
both of them and contains polyfills for a few Object.*.

Implementation is that there is now a `polyfill-wrapper-start.js` which
has the outer most closure, then the two polyfills (after being
    compiled), then `polyfill-wrapper-end.js`.
