+++ Gruntfile.js
@@ -19,21 +19,32 @@ module.exports = function (grunt) {
         'lib/system.js'
       ]
     },
-    esnext: {
+    concat: {
       dist: {
         src: [
           'node_modules/when/es6-shim/Promise.js',
-          'src/loader.js',
-          'src/system.js'
+          'src/polyfill-wrapper-start.js',
+          'dist/<%= pkg.name %>.js',
+          'src/polyfill-wrapper-end.js'
         ],
         dest: 'dist/<%= pkg.name %>.js'
       },
       polyfillOnly: {
         src: [
+          'src/polyfill-wrapper-start.js',
+          'dist/<%= pkg.name %>.js',
+          'src/polyfill-wrapper-end.js'
+        ],
+        dest: 'dist/<%= pkg.name %>-sans-promises.js'
+      }
+    },
+    esnext: {
+      dist: {
+        src: [
           'src/loader.js',
           'src/system.js'
         ],
-        dest: 'dist/<%= pkg.name %>-sans-promises.js'
+        dest: 'dist/<%= pkg.name %>.js'
       }
     },
     'string-replace': {
@@ -46,7 +57,7 @@ module.exports = function (grunt) {
             pattern: 'var $__Object$getPrototypeOf = Object.getPrototypeOf;\n' +
               'var $__Object$defineProperty = Object.defineProperty;\n' +
               'var $__Object$create = Object.create;',
-            replacement: "<%= grunt.file.read('src/object_polyfills.js') %>"
+            replacement: ''
           }]
 
         }
@@ -76,8 +87,10 @@ module.exports = function (grunt) {
   grunt.loadNpmTasks('grunt-contrib-jshint');
   grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-esnext');
+  grunt.loadNpmTasks('grunt-contrib-concat');
   grunt.loadNpmTasks('grunt-string-replace');
 
   grunt.registerTask('lint', ['jshint']);
-  grunt.registerTask('default', [/*'jshint', */'esnext', 'string-replace', 'uglify']);
+  grunt.registerTask('default', [/*'jshint', */'esnext', 'string-replace', 
+                     'concat', 'uglify']);
 };
