+++ package.json
@@ -23,6 +23,7 @@
   ],
   "devDependencies": {
     "grunt": "~0.4.1",
+    "grunt-contrib-concat": "^0.5.0",
     "grunt-contrib-jshint": "~0.6.0",
     "grunt-contrib-uglify": "~0.2.2",
     "grunt-esnext": "0.0.3",
