+++ src/polyfill-wrapper-end.js
@@ -1,0 +1,2 @@
+
+})(typeof global !== 'undefined' ? global : this);
